integrate a sentence handler for the tests
